Ref: Created /list and moved files inside
